requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
mysqlclient==2.2.0

# HTTP Requests
requests==2.31.0

# Security & CORS
django-cors-headers==4.3.1
argon2-cffi==23.1.0

# Environment
python-decouple==3.8
python-dotenv==1.0.1

# Utilities
pytz==2023.3
Pillow==10.1.0
PyYAML==6.0.1
python-dateutil==2.8.2

# Data Science & Machine Learning (for recommender)
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Task Queue
django-q2==1.6.2

# Production Server
gunicorn==21.2.0 # Linux
waitress==2.1.2 # Windows
whitenoise==6.6.0

# Development (comment out in production)
# django-debug-toolbar==4.2.0
# pytest==7.4.3
# pytest-django==4.7.0